Fix a bug where running in Junit would not load all classes
When IntelliJ calls Junit platform on a JPMS-enabled project,
it appears to put the target/classes and target/test-classes into
the jdk.module.path JVM variable rather than the java.class.path one.
This means that all other dependencies would get loaded, but not
the classes that a developer would be directly testing in their
projects. The fix for this is to add the jdk.module.path to the
StandardLocation.MODULE_PATH location. The compiler will still not
detect these classes correctly unless we also add the MODULE_PATH
paths we find into the StandardLocation.CLASS_PATH as well, so we
also duplicate that logic.