Disable ECJ tests for modules after more investigation work.
It appears the file management for modules in ECJ is totally bypassing the mechanism
we implement to read files and query the compilation unit file system. Instead, it is
using org.eclipse.jdt.internal.compiler.batch.FileSystem#initializeKnownFileNames and
using java.io.File, which will not work with non-default file systems. As for why this
works without modules, I have no idea.