JctFileManagerImpl: fix module inference logic
When getLocationForModule(Location, JavaFileObject) is invoked, there
is an expectation by some APIs that the method will return null if passed
a module-oriented/output-oriented location along with a file object
that is not associated with a module. The current implementation in JCT
raised an exception in this scenario instead, causing some libraries
such as manifold-systems to fail unexpectedly. This has been amended to
return null instead.