Versioning strategies in Junit support. (#274)
Implement ability to change default versioning strategy in JUnit
support.

If the user wishes to iterate across sources/targets instead of releases
for the JUnit annotations, there is now an attribute that can be
overridden to support that.

This is a breaking change in the JUnit API, but since this is not yet a fully stable release,
I am leaving the versioning as-is for this change. This only affects custom compiler
implementations with custom JUnit parameters anyway. For 99.99% of use
cases, this will not affect anything.