Deprecate all platform class path features for v1.0.0
Platform class path functionality in the JVM was historically used before JPMS introduced
system modules for the standard library. In lieu of that, the platform class path functionality
is mostly deprecated, and has been reported to not work nicely in tests. Therefore,
I am deprecating this with immediate effect for removal in v1.0.0.