Add ability to override debugger info flags/parameter info flags
Add the ability to set the debugger info flags explicitly, and default to
enabling said flags by default to improve the diagnostic capabilities of
generated class files.

Adds the ability to override the default parameter info inclusion setting
on the compiler, and defaults to including by default to enrich further
reflection assertions.