GH-721: Refactor all instances of methods taking 'T, T...' as an argument
This removes all APIs that consume an element and a varargs array of the same type
as this has been deemed to be clunky and somewhat confusing. The original intent was
to communicate that at least one element should be used.

The APIs have been replaced with a version that consumes a single varargs array
instead, which allows passing in an array directly.

Passing an empty array will now raise an `InvalidArgumentException` in places where it makes
no sense, which should continue to assist users in writing meaningful test cases.

**This is a breaking API change!**